+++ pkg/kv/kvserver/rangefeed/BUILD.bazel
@@ -8,13 +8,17 @@ go_library(
         "event_size.go",
         "filter.go",
         "metrics.go",
+        "nonbuffered_registration.go",
         "processor.go",
         "processor_test_helper.go",
+        "registration.go",
         "registry.go",
         "registry_test_helper.go",
         "resolved_timestamp.go",
         "scheduled_processor.go",
         "scheduler.go",
+        "stream_event_queue.go",
+        "stream_event_queue_helper.go",
         "stream_muxer.go",
         "stream_muxer_helper.go",
         "task.go",
